pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xmag"
version = "0.1.0"
description = "Convert X article URLs into three-column LaTeX PDFs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "typer>=0.12,<1",
  "playwright>=1.50,<2",
  "httpx>=0.27,<1",
  "jinja2>=3.1,<4",
  "pydantic>=2.8,<3",
  "python-dateutil>=2.9,<3",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9",
  "pytest-cov>=5,<6",
  "ruff>=0.6,<1",
  "mypy>=1.11,<2",
  "types-python-dateutil>=2.9.0",
]

[project.scripts]
xmag = "xmag.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
xmag = ["templates/*.tex.j2"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = true
packages = ["xmag"]
